Refactoring changed file names
