+++ src/unstract/sdk/utils/tool_utils.py
@@ -10,7 +10,7 @@ import magic
 
 from unstract.sdk.exceptions import FileStorageError
 from unstract.sdk.file_storage import FileStorage, FileStorageProvider
-from unstract.sdk.file_storage.fs_shared_temporary import SharedTemporaryFileStorage
+from unstract.sdk.file_storage.shared_temporary import SharedTemporaryFileStorage
 
 logger = logging.getLogger(__name__)
