+++ tests/test_fs_permanent.py
@@ -6,7 +6,7 @@ import pytest
 from dotenv import load_dotenv
 
 from unstract.sdk.file_storage import FileStorageProvider
-from unstract.sdk.file_storage.fs_permanent import PermanentFileStorage
+from unstract.sdk.file_storage.permanent import PermanentFileStorage
 
 load_dotenv()
